Split index to files
